+++ AmazonChimeSDK/AmazonChimeSDK/audiovideo/AudioVideoConfiguration.swift
@@ -16,6 +16,39 @@ import Foundation
     public let enableAudioRedundancy: Bool
     public let videoMaxResolution: VideoResolution
 
+    // These convenience initializers exist so that Objective-C code can initialize this class without specifying all the parameters.
+    convenience override public init() {
+        self.init()
+    }
+
+    convenience public init(audioMode: AudioMode) {
+        self.init(audioMode: audioMode)
+    }
+
+    convenience public init(audioDeviceCapabilities: AudioDeviceCapabilities) {
+        self.init(audioDeviceCapabilities: audioDeviceCapabilities)
+    }
+
+    convenience public init(callKitEnabled: Bool) {
+        self.init(callKitEnabled: callKitEnabled)
+    }
+
+    convenience public init(enableAudioRedundancy: Bool) {
+        self.init(enableAudioRedundancy: enableAudioRedundancy)
+    }
+
+    convenience public init(videoMaxResolution: VideoResolution) {
+        self.init(videoMaxResolution: videoMaxResolution)
+    }
+
+    convenience public init(audioMode: AudioMode, callKitEnabled: Bool) {
+        self.init(audioMode: audioMode, callKitEnabled: callKitEnabled)
+    }
+
+    convenience public init(audioMode: AudioMode, callKitEnabled: Bool, enableAudioRedundancy: Bool) {
+        self.init(audioMode: audioMode, callKitEnabled: callKitEnabled, enableAudioRedundancy: enableAudioRedundancy)
+    }
+
     public init(audioMode: AudioMode = .stereo48K,
                 audioDeviceCapabilities: AudioDeviceCapabilities = .inputAndOutput,
                 callKitEnabled: Bool = false,