+++ AmazonChimeSDK/AmazonChimeSDK/audiovideo/audio/AudioMode.swift
@@ -19,6 +19,9 @@ import Foundation
     /// The stereo audio mode with two audio channels for speaker, and single audio channel for microphone, both with 48KHz sampling rate.
     case stereo48K = 3
 
+    // Value 4 is reserved for the obsolete nodevice case, which is replaced by AudioDeviceCapabilities.none
+    // case nodevice = 4
+
     public var description: String {
         switch self {
         case .mono16K: